requirements.txt
Flask
flask-sqlalchemy
flask-migrate
flask-wtf
Markdown
gunicorn
python-dotenv